Collapse converter
Make it simpler, and do not use Plexus DI lifecycle interfaces. This
now uses Mojo configurators only (that still deeply depends on Plexus
internals, but we do not go around).
